// src/api/schemas/getIdiomaAgentes200DataItem.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAC API
 * Api para uso interno WAC.
 * OpenAPI spec version: 1.0.0
 */

export type GetIdiomaAgentes200DataItem = {
  id?: string;
  email?: string;
  pais?: string;
  nombre?: string;
  apellido?: string;
  telefono?: string;
  xxx_tipo?: string;
  password_ini?: string;
  ultimo_login?: string;
  imagen?: string;
  distribuidor?: string;
  comision?: string;
  rol?: string;
  email_registro?: string;
  email_bienvenida?: string;
  tipo_pago?: string;
  ts_creacion?: string;
  idioma?: string;
  qr?: string;
  recibir_correos_renovaciones?: string;
  correo_renovaciones_alternativo?: string;
  status?: string;
  codigo?: string;
  ultimo_login_fecha?: string;
  email_registro_fecha?: string;
  email_bienvenida_fecha?: string;
  fecha_creacion?: string;
};